.git/safe is opt-in
The idea behind the .git/safe is that I run it manually and explicitly
after I trust the directoy and the team behind the project.

Creating it as part of the setup script defeats the point of that. Doing
so without explcitly telling the user comes across as sneaky, perhaps.
(I do expect all devs to read setup scripts before running them -- but I
also know that they do not!)

Switch this script to inform the user that they have no `.git/safe`
directory but might like to make one.